+++ examples/namespace_within_module.js
@@ -1,18 +1,29 @@
-var myapp = $module(function () {
+var myapp = $module(function (myapp) {
 
-  $namespace('visual.shared');
+  $namespace('myapp.visual.common');
 
-  visual.shared.Element = $class({
+  myapp.visual.common.Button = $class({
     // Code goes here....
+  });
+
+  // Alternativlu you can write it in such manner
+  $namespace('myapp.visual.common', {
+
+    Button: $class({
+      // Code goes here....
+    })
+
   });
 
 });
 
 (function () {
 
+  'use strict';
+
   // In other scope local variable, which will point to the desired 
   // class, can be created 
 
-  var Element = myapp.visual.shared.Element;
+  var button = myapp.visual.common.Button;
 
 }());