+++ src/class4js.js
@@ -1,1 +1,1 @@
-exports.version = '1.10.1';
+exports.version = '1.10.3';