+++ tests/browser/scripts/namespace_within_module.js
@@ -1,0 +1,13 @@
+$run("Namespace within module", function () {
+
+  var myapp = $module(function (myapp, global) {
+
+    $namespace('myapp.visual.common');
+
+    $assert(myapp.visual.common);
+
+    $complete("Namespace within module");
+
+  });
+
+});