+++ tests/node/fields.js
@@ -4,6 +4,7 @@ require('../../lib/class4js.js');
 
 var Person = $class({
   __age: 30,
+  __comment: null,
   __construct__: function (name) {
     this.__name = name;
   },
@@ -16,13 +17,22 @@ var Person = $class({
   setGender: function (gender) {
     // It's to late to initialize a class field:
     this.__gender = gender;
+  },
+  getComment: function () {
+    return this.__comment;
+  },
+  setComment: function (comment) {
+    this.__comment = comment;
   }
 });
 
 var person = new Person('John Smith');
 
+person.setComment("Person's name is John");
+
 console.assert(person.getName() == 'John Smith');
 console.assert(person.getAge() == 30);
+console.assert(person.getComment() == "Person's name is John");
 
 var errorRaised = false;
 try {